"payment"

+++ model/delete_itemcart.php
@@ -1,8 +1,10 @@
 <?php
     include 'connect.php';
+    session_start();
+    $accountId = $_SESSION['accountId'];
     if(isset($_GET['item_id'])) {
         $item_id=$_GET['item_id'];
-        $sql="DELETE FROM cart WHERE cItem_id = $item_id";
+        $sql="DELETE FROM cart WHERE cItem_id = $item_id AND cUser_id = $accountId";
         $result = $conn->query($sql);
         if($result === TRUE) {
             echo "<script>alert('Deleted!')</script>";
